pyproject.toml
[project]
name = "nelow-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0",
    "langchain>=0.3.23",
    "openai>=1.75.0",
    "pandas>=2.2.3",
    "pdfplumber>=0.11.6",
    "pypdf2>=3.0.1",
    "sentence-transformers>=4.1.0",
]
